+++ web-app/asset-manager-frontend/js/main.js
@@ -1,6 +1,6 @@
 console.log('MAIN.JS: Entry point (v7.00)');
 import { showView } from './utils.js?v=6.60';
-import { renderDashboard, setupDashboard, setupDashboardFormHandlers, renderSidebarTree, editAsset } from './dashboard.js?v=7.00';
+import { renderDashboard, setupDashboard, setupDashboardFormHandlers, renderSidebarTree, editAsset } from './dashboard.js?v=7.02';
 // import { initScannerView } from './networkScanner.js?v=5.50';
 import { renderItAssets } from './itAssets.js?v=6.60';
 import { setupAuth, checkSession } from './auth.js?v=7.01';